requirements.txt
fastapi
uvicorn
pandas
pydantic-settings
python-dotenv
beautifulsoup4
requests